Remove pinned dependency on scikit-learn
Currently, I get an error when trying to install with this setup.py
in my environment (Debian testing, Python 3.7.3), using

pip install -e .

Apparently the pinned dependency on scikit-learn pulls an uninstallable
version of scikit (or maybe numpy, I'm not sure). Removing the pinned
dependency makes installation work. So, if that specific version of
scikit-learn is not needed, I suggest to just use this patch, and
release a new version to pypi.